avoid directly import extra dependencies
