+++ autoPyTorch/evaluation/abstract_evaluator.py
@@ -19,7 +19,11 @@ from smac.tae import StatusType
 import autoPyTorch.pipeline.image_classification
 import autoPyTorch.pipeline.tabular_classification
 import autoPyTorch.pipeline.tabular_regression
-import autoPyTorch.pipeline.time_series_forecasting
+try:
+    import autoPyTorch.pipeline.time_series_forecasting
+    forecasting_dependencies_installed = True
+except ModuleNotFoundError:
+    forecasting_dependencies_installed = False
 import autoPyTorch.pipeline.traditional_tabular_classification
 import autoPyTorch.pipeline.traditional_tabular_regression
 from autoPyTorch.automl_common.common.utils.backend import Backend
@@ -27,6 +31,7 @@ from autoPyTorch.constants import (
     CLASSIFICATION_TASKS,
     FORECASTING_BUDGET_TYPE,
     FORECASTING_TASKS,
+    ForecastingDependenciesNotInstalledError,
     IMAGE_TASKS,
     MULTICLASS,
     REGRESSION_TASKS,
@@ -38,12 +43,16 @@ from autoPyTorch.datasets.base_dataset import (
     BaseDataset,
     BaseDatasetPropertiesType
 )
-from autoPyTorch.datasets.time_series_dataset import TimeSeriesSequence
 from autoPyTorch.evaluation.utils import (
     DisableFileOutputParameters,
     VotingRegressorWrapper,
     convert_multioutput_multiclass_to_multilabel
 )
+try:
+    from autoPyTorch.evaluation.utils_extra import DummyTimeSeriesForecastingPipeline
+    forecasting_dependencies_installed = True
+except ModuleNotFoundError:
+    forecasting_dependencies_installed = False
 from autoPyTorch.pipeline.base_pipeline import BasePipeline
 from autoPyTorch.pipeline.components.training.metrics.base import autoPyTorchMetric
 from autoPyTorch.pipeline.components.training.metrics.utils import (
@@ -314,61 +323,6 @@ class DummyRegressionPipeline(DummyRegressor):
                 'runtime': 1}
 
 
-class DummyTimeSeriesForecastingPipeline(DummyClassificationPipeline):
-    """
-    A wrapper class that holds a pipeline for dummy forecasting. For each series, it simply repeats the last element
-    in the training series
-
-
-    Attributes:
-        random_state (Optional[Union[int, np.random.RandomState]]):
-            Object that contains a seed and allows for reproducible results
-        init_params  (Optional[Dict]):
-            An optional dictionary that is passed to the pipeline's steps. It complies
-            a similar function as the kwargs
-        n_prediction_steps (int):
-            forecasting horizon
-    """
-    def __init__(self, config: Configuration,
-                 random_state: Optional[Union[int, np.random.RandomState]] = None,
-                 init_params: Optional[Dict] = None,
-                 n_prediction_steps: int = 1,
-                 ) -> None:
-        super(DummyTimeSeriesForecastingPipeline, self).__init__(config, random_state, init_params)
-        self.n_prediction_steps = n_prediction_steps
-
-    def fit(self, X: Dict[str, Any], y: Any,
-            sample_weight: Optional[np.ndarray] = None) -> object:
-        self.n_prediction_steps = X['dataset_properties']['n_prediction_steps']
-        y_train = subsampler(X['y_train'], X['train_indices'])
-        return DummyClassifier.fit(self, np.ones((y_train.shape[0], 1)), y_train, sample_weight)
-
-    def _generate_dummy_forecasting(self, X: List[Union[TimeSeriesSequence, np.ndarray]]) -> List:
-        if isinstance(X[0], TimeSeriesSequence):
-            X_tail = [x.get_target_values(-1) for x in X]
-        else:
-            X_tail = [x[-1] for x in X]
-        return X_tail
-
-    def predict_proba(self, X: Union[np.ndarray, pd.DataFrame],
-                      batch_size: int = 1000) -> np.ndarray:
-        X_tail = self._generate_dummy_forecasting(X)
-        return np.tile(X_tail, (1, self.n_prediction_steps)).astype(np.float32).flatten()
-
-    def predict(self, X: Union[np.ndarray, pd.DataFrame],
-                batch_size: int = 1000) -> np.ndarray:
-        X_tail = np.asarray(self._generate_dummy_forecasting(X))
-        if X_tail.ndim == 1:
-            X_tail = np.expand_dims(X_tail, -1)
-        return np.tile(X_tail, (1, self.n_prediction_steps)).astype(np.float32).flatten()
-
-    @staticmethod
-    def get_default_pipeline_options() -> Dict[str, Any]:
-        return {'budget_type': 'epochs',
-                'epochs': 1,
-                'runtime': 1}
-
-
 def fit_and_suppress_warnings(logger: PicklableClientLogger, pipeline: BaseEstimator,
                               X: Dict[str, Any], y: Any
                               ) -> BaseEstimator:
@@ -543,6 +497,8 @@ class AbstractEvaluator(object):
             self.predict_function = self._predict_proba
         elif self.task_type in FORECASTING_TASKS:
             if isinstance(self.configuration, int):
+                if not forecasting_dependencies_installed:
+                    raise ForecastingDependenciesNotInstalledError
                 self.pipeline_class = DummyTimeSeriesForecastingPipeline
             elif isinstance(self.configuration, str):
                 raise ValueError("Only tabular classifications tasks "
